refactor: Move 'center on location' button from Home to Navbar component. Hoist globeEl and location variables to App in order to pass as props to Navbar
