+++ frontend/src/components/Navbar.jsx
@@ -1,9 +1,19 @@
 import { ConnectButton } from "@rainbow-me/rainbowkit";
 import { useAccount } from "wagmi";
 
-export default function Navbar() {
+export default function Navbar({location, globeEl}) {
   const { data } = useAccount();
   const pathName = window.location.pathname;
+
+  const centerOnLocation = () => {
+    const MAP_CENTER = {
+      lat: location?.coordinates?.lat,
+      lng: location?.coordinates?.lon,
+      altitude: 2.5,
+    };
+    const ROTATION_SPEED = 500;
+    globeEl.current.pointOfView(MAP_CENTER, ROTATION_SPEED);
+  };
 
   return (
     <>
@@ -54,6 +64,30 @@ export default function Navbar() {
             showBalance={false}
           />
         </div>
+        <button
+            class="btn btn-round  p-2 py-1 m-3"
+            onClick={centerOnLocation}
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              class="h-6 w-6"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+              stroke-width="2"
+            >
+              <path
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                d="M17.657 16.657L13.414 20.9a1.998 1.998 0 01-2.827 0l-4.244-4.243a8 8 0 1111.314 0z"
+              />
+              <path
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                d="M15 11a3 3 0 11-6 0 3 3 0 016 0z"
+              />
+            </svg>
+          </button>
       </div>
     </>
   );
