+++ frontend/src/views/Home.jsx
@@ -6,13 +6,10 @@ import Modal from "../components/Modal";
 import TransactionProgress from "../components/TransactionProgress";
 
 import useLoading from "../hooks/useLoading";
-import useGeolocation from "../hooks/useGeolocation";
 import { getMarkers } from "../helpers/getMarkers";
 import buyShopItem from "../helpers/buyShopItem";
 
-function Home() {
-  const globeEl = useRef();
-  const { location } = useGeolocation();
+function Home({globeEl, location}) {
   const [showModal, setShowModal] = useState(false);
   const [nft, setNft] = useState({});
   const [transactionInProgress, setTransactionInProgress] = useState(false);
@@ -56,16 +53,6 @@ function Home() {
     }, 200);
   };
 
-  const centerOnLocation = () => {
-    const MAP_CENTER = {
-      lat: location?.coordinates?.lat,
-      lng: location?.coordinates?.lon,
-      altitude: 2.5,
-    };
-    const ROTATION_SPEED = 500;
-    globeEl.current.pointOfView(MAP_CENTER, ROTATION_SPEED);
-  };
-
   return (
     <>
       {!isLoaded && (
@@ -86,30 +73,6 @@ function Home() {
             markers={markers.current}
             userLocation={location}
           />
-          <button
-            class="btn btn-round absolute bottom-0 right-0 m-3"
-            onClick={centerOnLocation}
-          >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              class="h-6 w-6"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-              stroke-width="2"
-            >
-              <path
-                stroke-linecap="round"
-                stroke-linejoin="round"
-                d="M17.657 16.657L13.414 20.9a1.998 1.998 0 01-2.827 0l-4.244-4.243a8 8 0 1111.314 0z"
-              />
-              <path
-                stroke-linecap="round"
-                stroke-linejoin="round"
-                d="M15 11a3 3 0 11-6 0 3 3 0 016 0z"
-              />
-            </svg>
-          </button>
         </>
       )}
       <Transition show={showModal}>
